oakd/requirements.txt
addict==2.4.0
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
ConfigArgParse==1.7.1
contourpy==1.3.2
cycler==0.12.1
dash==3.4.0
decorator==5.2.1
depthai==2.32.0.0
exceptiongroup==1.3.1
executing==2.2.1
fastjsonschema==2.21.2
Flask==3.1.2
fonttools==4.61.1
idna==3.11
importlib_metadata==8.7.1
ipython==8.38.0
ipywidgets==8.1.8
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
narwhals==2.15.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.2.6
open3d==0.19.0
opencv-python==4.13.0.90
packaging==26.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
plotly==6.5.2
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.3.2
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
retrying==1.4.2
rpds-py==0.30.0
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
stack-data==0.6.3
threadpoolctl==3.6.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
waitress==3.0.2
wcwidth==0.5.0
Werkzeug==3.1.5
widgetsnbextension==4.0.15
zipp==3.23.0
